+++ R/utils-checkmate.R
@@ -1,6 +1,15 @@
 checkHasNames <- function(x, n, type = "unique") checkmate::checkNames(names(x), must.include = n, type = type)
 assertHasNames <- checkmate::makeAssertionFunction(checkHasNames)
 
+checkRange <- function(x, null.ok = FALSE)
+{
+    ret <- checkmate::checkNumeric(x, any.missing = FALSE, finite = TRUE, len = 2, null.ok = null.ok)
+    if (isTRUE(ret) && !is.null(x) && x[2] != -1 && x[1] > x[2])
+        ret <- paste0("lower range (", x[1], ") higher than upper (", x[2], ")")
+    return(ret)
+}
+assertRange <- checkmate::makeAssertionFunction(checkRange)
+
 checkChoiceSilent <- function(x, ch)
 {
     ret <- checkmate::checkString(x, min.chars = 1)